Move the component and fix lint error
