+++ src/components/CircleAvatar/CircleAvatarWithIcon.stories.tsx
