+++ src/components/CircleAvatar/type.ts
