+++ src/stories/components/CUActivityTable/CUActivityItem.tsx
@@ -5,10 +5,10 @@ import { DateTime } from 'luxon';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useMemo } from 'react';
+import CircleAvatar from '@/components/CircleAvatar/CircleAvatar';
 import lightTheme from '../../../../styles/theme/themes';
 import { useThemeContext } from '../../../core/context/ThemeContext';
 import { ButtonType } from '../../../core/enums/buttonTypeEnum';
-import { CircleAvatar } from '../CircleAvatar/CircleAvatar';
 import { CustomButton } from '../CustomButton/CustomButton';
 import { getCorrectCodeFromActivity, getResourceType } from './utils/helpers';
 import type { Activity } from './ActivityTable';
