+++ src/stories/components/CardInfoMember/CardInfoMember.tsx
@@ -1,10 +1,10 @@
 import { Box, Card, CardContent, CardHeader, Divider, Typography, styled } from '@mui/material';
 import { DateTime } from 'luxon';
 import React from 'react';
+import CircleAvatar from '@/components/CircleAvatar/CircleAvatar';
 import { getContributorCommitment, getLinksFromContributor } from '../../../core/businessLogic/coreUnitAbout';
 import { useThemeContext } from '../../../core/context/ThemeContext';
 import { getColorJobPosition } from '../../../core/utils/colors';
-import { CircleAvatar } from '../CircleAvatar/CircleAvatar';
 import { CuTableColumnLinks } from '../CuTableColumnLinks/CuTableColumnLinks';
 import { CustomPopover } from '../CustomPopover/CustomPopover';
 import type { ContributorCommitment } from '@ses/core/models/interfaces/contributor';
